chore: account for types moved around
